requirements.txt
flask==3.0.3
google-generativeai==0.5.2
python-dotenv==1.0.1
gunicorn==21.2.0
uvicorn==0.29.0 # Added for uvicorn worker class
